pyproject.toml
[project]
name = "retro-rl-agents"
version = "0.1.0"
description = "Train RL agents with stable-retro (mainly focused on retro fighting games)"
readme = "README.md"
requires-python = "~=3.12.0"
dependencies = [
    "imitation>=1.0.0",
    "opencv-python>=4.13.0.92",
    "pyyaml>=6.0.3",
    "setuptools==81.0.0",
    "stable-baselines3[extra]>=2.7.1",
    "tbparse>=0.0.9",
    "tensorboard>=2.20.0",
]

[project.scripts]
agents = "retro_rl_agents.main:main"

[tool.uv]
required-version = ">=0.8.17"

[tool.uv.workspace]
exclude = [
    "lib/stable-retro",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.4",
    "sphinx>=9.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/retro_rl_agents"]

[tool.ruff]
target-version = "py312"
line-length = 80
indent-width = 4
exclude = ["fusenet.py"]

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
